+++ src/main/java/by/novikgrits/webapp/model/repository/LotRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LotRepository extends CrudRepository<Lot, Long> {
+public interface LotRepository extends CrudRepository<Lot, Integer> {
 }
