+++ src/main/java/by/novikgrits/webapp/model/repository/UserRoleRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRoleRepository extends CrudRepository<UserRole, Long> {
+public interface UserRoleRepository extends CrudRepository<UserRole, Integer> {
 }
